refactor : hamc, jwt Authentication 모두 CustomAuthentication 을 상속 받도록 변경
