+++ src/main/kotlin/com/flab/inqueue/security/common/CustomAuthentication.kt
@@ -1,0 +1,41 @@
+package com.flab.inqueue.security.common
+
+import org.springframework.security.core.Authentication
+import org.springframework.security.core.GrantedAuthority
+
+
+abstract class CustomAuthentication(
+    private val name: String?,
+    private val principal : Any?,
+    private var isAuthenticated : Boolean,
+    private val authorities: MutableCollection<out GrantedAuthority> = mutableListOf()
+) : Authentication {
+
+    override fun getName(): String? {
+        return name
+    }
+
+    override fun getAuthorities(): MutableCollection<out GrantedAuthority> {
+        return this.authorities
+    }
+
+    override fun getCredentials(): Any? {
+        return null
+    }
+
+    override fun getDetails(): Any? {
+        return null
+    }
+
+    override fun getPrincipal(): Any? {
+        return this.principal
+    }
+
+    override fun isAuthenticated(): Boolean {
+        return isAuthenticated
+    }
+
+    override fun setAuthenticated(isAuthenticated: Boolean) {
+        this.isAuthenticated = isAuthenticated
+    }
+}
