+++ src/main/kotlin/com/flab/inqueue/security/hmacsinature/HmacAuthenticationToken.kt
@@ -1,17 +1,14 @@
 package com.flab.inqueue.security.hmacsinature
 
-import org.springframework.security.core.Authentication
+import com.flab.inqueue.security.common.CustomAuthentication
 import org.springframework.security.core.GrantedAuthority
 
 class HmacAuthenticationToken(
     val clientId: String? = null,
     val signature: String? = null,
     val payload: String? = null,
-    private val authorities: MutableCollection<out GrantedAuthority> = mutableListOf()
-) : Authentication {
-
-    private var principal = clientId
-    private var isAuthenticated = false
+    authorities: MutableCollection<out GrantedAuthority> = mutableListOf()
+) : CustomAuthentication(clientId, clientId, false, authorities) {
 
     companion object {
         @JvmStatic
@@ -24,33 +21,5 @@ class HmacAuthenticationToken(
                 authorities = authorities
             ).apply { isAuthenticated = true }
         }
-    }
-
-    override fun getName(): String {
-        return this.clientId!!
-    }
-
-    override fun getAuthorities(): MutableCollection<out GrantedAuthority> {
-        return this.authorities
-    }
-
-    override fun getCredentials(): Any? {
-        return null
-    }
-
-    override fun getDetails(): Any? {
-        return null
-    }
-
-    override fun getPrincipal(): Any? {
-        return principal
-    }
-
-    override fun isAuthenticated(): Boolean {
-        return this.isAuthenticated
-    }
-
-    override fun setAuthenticated(isAuthenticated: Boolean) {
-        this.isAuthenticated = isAuthenticated
     }
 }
