+++ src/main/kotlin/com/flab/inqueue/security/jwt/JwtAuthenticationToken.kt
@@ -1,15 +1,13 @@
 package com.flab.inqueue.security.jwt
 
-import org.springframework.security.core.Authentication
+import com.flab.inqueue.security.common.CustomAuthentication
 import org.springframework.security.core.GrantedAuthority
 
 class JwtAuthenticationToken(
-    private val userId: String? = null,
+    userId: String? = null,
     val jwtToken: String? = null,
-    private val authorities: MutableCollection<out GrantedAuthority> = mutableListOf()
-) : Authentication {
-
-    private var isAuthenticated = false
+    authorities: MutableCollection<out GrantedAuthority> = mutableListOf()
+) : CustomAuthentication(userId, userId, false, authorities) {
 
     companion object {
         @JvmStatic
@@ -20,33 +18,5 @@ class JwtAuthenticationToken(
                 userId = userId
             ).apply { isAuthenticated = true }
         }
-    }
-
-    override fun getName(): String? {
-        return this.userId
-    }
-
-    override fun getAuthorities(): MutableCollection<out GrantedAuthority> {
-        return authorities
-    }
-
-    override fun getCredentials(): Any? {
-        return null
-    }
-
-    override fun getDetails(): Any? {
-        return null
-    }
-
-    override fun getPrincipal(): Any? {
-        return userId
-    }
-
-    override fun isAuthenticated(): Boolean {
-        return this.isAuthenticated
-    }
-
-    override fun setAuthenticated(isAuthenticated: Boolean) {
-        this.isAuthenticated = isAuthenticated
     }
 }
